move remove into its own file
